// src/lib/api/schemas/newHireResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { NewHireResponseInvitedAt } from './newHireResponseInvitedAt';
import type { NewHireResponseStartedAt } from './newHireResponseStartedAt';
import type { NewHireResponseCompletedAt } from './newHireResponseCompletedAt';
import type { NewHireResponseSessionToken } from './newHireResponseSessionToken';
import type { NewHireResponseSessionTokenExpiresAt } from './newHireResponseSessionTokenExpiresAt';
import type { NewHireResponseUpdatedAt } from './newHireResponseUpdatedAt';

export interface NewHireResponse {
  email: string;
  first_name: string;
  last_name: string;
  flow_id: string;
  id: string;
  company_id: string;
  status: string;
  invited_at?: NewHireResponseInvitedAt;
  started_at?: NewHireResponseStartedAt;
  completed_at?: NewHireResponseCompletedAt;
  session_token?: NewHireResponseSessionToken;
  session_token_expires_at?: NewHireResponseSessionTokenExpiresAt;
  created_at: string;
  updated_at?: NewHireResponseUpdatedAt;
}